+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## v2.8.0-dev
+## v2.8.0
 
 * Enhancements
     * Added logging to the Login, OnetimePass, ConfirmEmail and ResetPassword modules
-    * Made Login, ConfirmEmail and ResetPassword modules more configurable
+    * Made ConfirmEmail and ResetPassword modules more configurable
 
 ## v2.7.0